refactor: tournaments/layout.tsx is now built server side, and a new component has been added for the top container
