+++ src/app/tournaments/layout.module.scss
@@ -1,37 +1,6 @@
 @import '@/variables.scss';
 
 .tournamentLayout {
-  .topContainer {
-    background-color: $primary-background;
-    padding: 100px 0 100px 0;
-
-    .top {
-      width: 65%;
-      margin: 0 auto;
-
-      .text {
-        font-size: 1.25rem;
-        text-align: center;
-        font-weight: 500;
-        width: 80%;
-        margin: 0 auto;
-        color: $blue-color;
-      }
-
-      .buttons {
-        display: flex;
-        justify-content: center;
-        column-gap: 92px;
-        padding: 50px 0;
-
-        .button {
-          width: 328px;
-          height: 66px
-        }
-      }
-    }
-  }
-
   .mainContent {
     position: relative;
     padding-bottom: 3rem;
